# bot/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Bot Service API
  version: 1.0.0
  description: |
    REST API сервиса bot для управления клиентами, подписками и приема уведомлений от scrapper.
servers:
  - url: http://localhost:8080
    description: Local environment

tags:
  - name: Clients
    description: Регистрация, вход и удаление клиента
  - name: Links
    description: Управление подписками на ссылки
  - name: Updates
    description: Входящие уведомления от scrapper
  - name: Notifications
    description: Чтение и управление статусом уведомлений
  - name: PushTokens
    description: Регистрация и деактивация FCM device token

paths:
  /api/v1/clients:
    post:
      tags: [Clients]
      operationId: registerClient
      summary: Зарегистрировать клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCredentialsRequest'
      responses:
        '200':
          description: Клиент успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Clients]
      operationId: unregisterClient
      summary: Удалить клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCredentialsRequest'
      responses:
        '200':
          description: Клиент успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/clients/login:
    post:
      tags: [Clients]
      operationId: loginClient
      summary: Выполнить вход клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCredentialsRequest'
      responses:
        '200':
          description: Вход выполнен успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/links:
    get:
      tags: [Links]
      operationId: getLinks
      summary: Получить список отслеживаемых ссылок
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      responses:
        '200':
          description: Список ссылок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Links]
      operationId: trackLink
      summary: Подписаться на ссылку
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLinkRequest'
      responses:
        '200':
          description: Подписка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Links]
      operationId: untrackLink
      summary: Отписаться от ссылки
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveLinkRequest'
      responses:
        '200':
          description: Подписка удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /updates:
    post:
      tags: [Updates]
      operationId: notifyLinkUpdate
      summary: Принять уведомление об обновлении ссылки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkUpdate'
      responses:
        '200':
          description: Уведомление принято
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/notifications:
    get:
      tags: [Notifications]
      operationId: getNotifications
      summary: Получить список уведомлений клиента
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: tags
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Список уведомлений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/notifications/unread-count:
    get:
      tags: [Notifications]
      operationId: getUnreadCount
      summary: Получить количество непрочитанных уведомлений
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      responses:
        '200':
          description: Количество непрочитанных уведомлений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/notifications/mark-read:
    post:
      tags: [Notifications]
      operationId: markNotificationsRead
      summary: Пометить уведомления как прочитанные
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkReadRequest'
      responses:
        '200':
          description: Результат операции mark read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkReadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/push-tokens:
    post:
      tags: [PushTokens]
      operationId: registerPushToken
      summary: Зарегистрировать или обновить FCM token
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenUpsertRequest'
      responses:
        '200':
          description: Токен активирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Превышен rate limit token API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /api/v1/push-tokens/current:
    delete:
      tags: [PushTokens]
      operationId: deactivatePushToken
      summary: Деактивировать текущий FCM token
      parameters:
        - $ref: '#/components/parameters/ClientLoginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenDeactivateRequest'
      responses:
        '200':
          description: Токен деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotApiMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Превышен rate limit token API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

components:
  parameters:
    ClientLoginHeader:
      name: Client-Login
      in: header
      required: true
      description: Логин клиента
      schema:
        type: string
        minLength: 1

  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Отсутствует внутренний auth-заголовок
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Неверный внутренний auth-секрет
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'

  schemas:
    ClientCredentialsRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          minLength: 1
          example: alice
        password:
          type: string
          minLength: 1
          example: qwerty123

    BotApiMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Операция выполнена успешно

    AddLinkRequest:
      type: object
      required: [link]
      properties:
        link:
          type: string
          format: uri
          example: https://github.com/spring-projects/spring-framework
        tags:
          type: array
          items:
            type: string
          uniqueItems: true
          nullable: true
          example: [backend, work]
        filters:
          type: array
          items:
            type: string
          uniqueItems: true
          nullable: true
          example: [author:octocat]

    RemoveLinkRequest:
      type: object
      required: [link]
      properties:
        link:
          type: string
          format: uri
          example: https://stackoverflow.com/questions/12345678/example

    LinkResponse:
      type: object
      required: [id, url]
      properties:
        id:
          type: integer
          format: int64
          example: 42
        url:
          type: string
          format: uri
          example: https://github.com/spring-projects/spring-framework
        tags:
          type: array
          items:
            type: string
          uniqueItems: true
          nullable: true
          example: [backend]
        filters:
          type: array
          items:
            type: string
          uniqueItems: true
          nullable: true
          example: [author:octocat]

    LinkUpdate:
      type: object
      required: [id, url, title, updateOwner, description, creationDate, clientLogins]
      properties:
        id:
          type: integer
          format: int64
          example: 101
        url:
          type: string
          format: uri
          example: https://github.com/spring-projects/spring-framework/pull/1
        eventUrl:
          type: string
          format: uri
          nullable: true
          example: https://github.com/spring-projects/spring-framework/pull/1
        title:
          type: string
          example: New issue created
        updateOwner:
          type: string
          example: octocat
        description:
          type: string
          example: Краткое описание обновления
        creationDate:
          type: string
          format: date-time
          example: '2026-04-05T10:00:00Z'
        clientLogins:
          type: array
          items:
            type: string
          example: [alice, bob]

    ApiErrorResponse:
      type: object
      required:
        - description
        - code
        - exceptionName
        - exceptionMessage
        - stacktrace
      properties:
        description:
          type: string
          example: Bad request
        code:
          type: string
          example: '400'
        exceptionName:
          type: string
          example: BadRequestException
        exceptionMessage:
          type: string
          example: Некорректные параметры запроса
        stacktrace:
          type: array
          items:
            type: string

    NotificationDto:
      type: object
      required: [id, linkId, url, title, updateOwner, description, creationDate, receivedAt, unread]
      properties:
        id:
          type: integer
          format: int64
          example: 10
        linkId:
          type: integer
          format: int64
          example: 101
        url:
          type: string
          format: uri
          example: https://github.com/org/repo/issues/1
        eventUrl:
          type: string
          format: uri
          nullable: true
          example: https://github.com/org/repo/issues/1
        title:
          type: string
          example: New issue comment
        updateOwner:
          type: string
          example: octocat
        description:
          type: string
          example: Added a new comment
        creationDate:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
        unread:
          type: boolean
          example: true

    NotificationListResponse:
      type: object
      required: [notifications, limit, offset]
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/NotificationDto'
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    UnreadCountResponse:
      type: object
      required: [unreadCount]
      properties:
        unreadCount:
          type: integer
          format: int64
          example: 3

    MarkReadRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
            format: int64
          uniqueItems: true
        markAll:
          type: boolean
          default: false

    MarkReadResponse:
      type: object
      required: [updatedCount]
      properties:
        updatedCount:
          type: integer
          format: int64
          example: 1

    PushTokenUpsertRequest:
      type: object
      required: [platform, token]
      properties:
        platform:
          type: string
          enum: [android]
          example: android
        token:
          type: string
          minLength: 16
          maxLength: 4096
          example: fcm-token-1234567890
        appVersion:
          type: string
          maxLength: 64
          nullable: true
          example: 1.4.2
        deviceId:
          type: string
          maxLength: 255
          nullable: true
          example: android-emulator-5554

    PushTokenDeactivateRequest:
      type: object
      required: [platform, token]
      properties:
        platform:
          type: string
          enum: [android]
          example: android
        token:
          type: string
          minLength: 16
          maxLength: 4096
          example: fcm-token-1234567890

    PushPayloadContract:
      type: object
      required: [event_id, title, content]
      description: Минимальный server-driven payload, отправляемый в FCM `data`.
      properties:
        event_id:
          type: string
          description: Уникальный id события в backend
        title:
          type: string
        content:
          type: string
        url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        source:
          type: string